Worked on Data Source Tab

+++ MachineLearning/src/java/edu/data/DataSet.java
@@ -53,7 +53,7 @@ public class DataSet
                     newAttributes.add(this.getFloat(attributes[i]));
                 }
                 patterns.add(
-                        Pattern.getPattern((Float[])newAttributes.toArray()));
+                        Pattern.getPattern((Double[])newAttributes.toArray()));
             }
             
         }
